+++ packages/toujou-location-finder/src/store/reducers/_popup-reducer.ts
@@ -16,18 +16,18 @@ const initialState = {
 export function popupReducer(state = initialState, action) {
   // eslint-disable-next-line default-case
   switch (action.type) {
-  case SET_POPUP_LOADING_START:
-    return { ...state, isLoading: true };
-  case SET_POPUP_LOADING_END:
-    return { ...state, isLoading: false };
-  case RESET_POPUP_FEATURE:
-    return { ...state, feature: null };
-  case SET_POPUP_FEATURE:
-    return { ...state, feature: action.payload };
-  case SET_POPUP_COORDINATES:
-    return { ...state, coordinates: action.payload };
-  case RESET_POPUP_COORDINATES:
-    return { ...state, coordinates: null };
+    case SET_POPUP_LOADING_START:
+      return { ...state, isLoading: true };
+    case SET_POPUP_LOADING_END:
+      return { ...state, isLoading: false };
+    case RESET_POPUP_FEATURE:
+      return { ...state, feature: null };
+    case SET_POPUP_FEATURE:
+      return { ...state, feature: action.payload };
+    case SET_POPUP_COORDINATES:
+      return { ...state, coordinates: action.payload };
+    case RESET_POPUP_COORDINATES:
+      return { ...state, coordinates: null };
   }
 
   return state;