+++ packages/toujou-poster-reveal/src/index.ts
@@ -1,1 +1,2 @@
 import './toujou-poster-reveal';
+import './poster-reveal-video-player';