+++ packages/toujou-poster-reveal/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the `@toujou/toujou-poster-reveal` package will be documented in this file
 
+## [1.1.0] - 2026-07-17
+- Add poster-reveal-video-player to the component
+- Update README and tests
+
 ## [1.0.0]
 - 🌱 Initial release
 - Add component, story and tests